+++ backend/app/models/__init__.py
@@ -1,0 +1,3 @@
+from app.models.base import Base
+from app.models.user_model import User
+from app.models.application_model import Application
