+++ backend/app/models/user_model.py
@@ -2,7 +2,7 @@ from pydantic import EmailStr
 from sqlalchemy import String
 from sqlalchemy.orm import Mapped, mapped_column
 
-from models.base import Base
+from app.models.base import Base
 
 
 class User(Base):
