+++ backend/app/schemas/user.py
@@ -1,0 +1,12 @@
+from pydantic import EmailStr
+from app.schemas.base import BaseSchema
+from app.schemas.application import Application
+
+
+class User(BaseSchema):
+    email: EmailStr
+
+
+class UserWithApplications(BaseSchema):
+    email: EmailStr
+    applications: list[Application] = []
