--- backend/models/__init__.py
@@ -1,3 +1,0 @@
-from models.base import Base
-from models.user_model import User
-from models.application_model import Application
